+++ src/Macros/MapToCollection.php
@@ -9,20 +9,15 @@ class MapToCollection
 {
     public function __invoke(): Closure
     {
-        return function (array $ary, bool $deep = false): Collection {
-            $closure = function (&$ary) use ($deep) {
-                if (is_array($ary)) {
-                    $ary = Collection::mapToCollection($ary, $deep);
-                }
+        return function (array $ary = [], bool $deep = false): Collection {
+            $ary = Collection::mapToCollectionFrom($ary, $deep);
+            $data =  Collection::mapToCollectionFrom($this->all(), $deep);
 
-                if ($deep && is_object($ary) && method_exists($ary, 'toArray')) {
-                    $ary = Collection::mapToCollection($ary->toArray(), $deep);
-                }
-            };
+            if ($ary->isNotEmpty()) {
+                return $data->merge([$ary]);
+            }
 
-            array_walk($ary, $closure);
-
-            return new Collection($ary);
+            return $data;
         };
     }
 }