+++ src/register.php
@@ -2,13 +2,8 @@
 
 use Illuminate\Support\Collection;
 
-$macros = [
-  'mapToCollection' => \Pulli\CollectionMacros\Macros\MapToCollection::class,
-  'recursiveToArray' => \Pulli\CollectionMacros\Macros\RecursiveToArray::class,
-];
-
-if (!function_exists('app')) {
-    Collection::make($macros)
+if (!class_exists('\Illuminate\Foundation\Application')) {
+    Collection::make(\Pulli\CollectionMacros\CollectionMacroServiceProvider::macros())
         ->reject(fn ($class, $macro) => Collection::hasMacro($macro))
         ->each(fn ($class, $macro) => Collection::macro($macro, (new $class())()));
 }