+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `collection-macros` will be documented in this file
 
+## 0.2.0 - 2023-12-18
+
+- Add Laravel ServiceProvider
+- Rewrite of mapToCollection / recursiveToArray as Collection functions
+- Introducing mapToCollectionFrom / recursiveToArrayFrom as static Collection functions
+
 ## 0.1.0 - 2023-09-12
 
 - Initial release